+++ smn-sdk-java/src/main/java/com/smn/request/sms/GetSmsMessageRequest.java
@@ -39,6 +39,7 @@ public class GetSmsMessageRequest extends AbstractRequest<GetSmsMessageResponse>
         }
 
         StringBuilder sb = new StringBuilder();
+        sb.append(getSmnServiceUrl());
         sb.append(Constants.URL_DELIMITER).append(Constants.V2).append(Constants.URL_DELIMITER)
                 .append(projectId).append(Constants.URL_DELIMITER).append(Constants.SMN_NOTIFICATIONS)
                 .append(Constants.URL_DELIMITER).append(Constants.SMN_SUB_PROTOCOL_SMS)
