+++ smn-sdk-java/src/main/java/com/smn/request/topic/CreateTopicRequest.java
@@ -51,6 +51,7 @@ public class CreateTopicRequest extends AbstractRequest<CreateTopicResponse> {
         }
 
         StringBuilder sb = new StringBuilder();
+        sb.append(getSmnServiceUrl());
         sb.append(Constants.URL_DELIMITER).append(Constants.V2).append(Constants.URL_DELIMITER)
                 .append(projectId).append(Constants.URL_DELIMITER).append(Constants.SMN_NOTIFICATIONS)
                 .append(Constants.URL_DELIMITER).append(Constants.TOPICS);
