+++ smn-sdk-java/src/main/java/com/smn/response/sms/ListSmsSignsResponse.java
@@ -45,66 +45,3 @@ public class ListSmsSignsResponse extends AbstractResponse {
         this.smsSigns = smsSigns;
     }
 }
-
-/**
- * sms sign info data
- *
- * @author zhangyx
- * @version 2.0.0
- */
-class SmsSignInfo {
-    @JsonProperty("sign_name")
-    private String signName;
-
-    @JsonProperty("create_time")
-    private String createTime;
-
-    @JsonProperty("sign_id")
-    private String signId;
-
-    @JsonProperty("reply")
-    private String reply;
-
-    @JsonProperty("status")
-    private int status;
-
-    public String getSignName() {
-        return signName;
-    }
-
-    public void setSignName(String signName) {
-        this.signName = signName;
-    }
-
-    public String getCreateTime() {
-        return createTime;
-    }
-
-    public void setCreateTime(String createTime) {
-        this.createTime = createTime;
-    }
-
-    public String getSignId() {
-        return signId;
-    }
-
-    public void setSignId(String signId) {
-        this.signId = signId;
-    }
-
-    public String getReply() {
-        return reply;
-    }
-
-    public void setReply(String reply) {
-        this.reply = reply;
-    }
-
-    public int getStatus() {
-        return status;
-    }
-
-    public void setStatus(int status) {
-        this.status = status;
-    }
-}
