+++ smn-sdk-java/src/main/java/com/smn/response/sms/SmsCallback.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017. Huawei Technologies Co., LTD. All rights reserved.
+ * Copyright (C) 2018. Huawei Technologies Co., LTD. All rights reserved.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of Apache License, Version 2.0.
@@ -12,30 +12,27 @@
 package com.smn.response.sms;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.smn.response.AbstractResponse;
-
-import java.util.List;
 
 /**
- * the response data of list Sms event
+ * Sms callback
  *
  * @author zhangyx
  * @version 2.0.0
  */
-public class ListSmsEventResponse extends AbstractResponse {
+public class SmsCallback {
 
-    @JsonProperty("callback")
-    private List<SmsCallback> callback;
+    @JsonProperty("event_type")
+    private String eventType;
 
     @JsonProperty("topic_urn")
     private String topicUrn;
 
-    public List<SmsCallback> getCallback() {
-        return callback;
+    public String getEventType() {
+        return eventType;
     }
 
-    public void setCallback(List<SmsCallback> callback) {
-        this.callback = callback;
+    public void setEventType(String eventType) {
+        this.eventType = eventType;
     }
 
     public String getTopicUrn() {
@@ -46,5 +43,3 @@ public class ListSmsEventResponse extends AbstractResponse {
         this.topicUrn = topicUrn;
     }
 }
-
-
