+++ smn-sdk-java/src/main/java/com/smn/response/sms/SmsSignInfo.java
@@ -1,0 +1,77 @@
+/*
+ * Copyright (C) 2018. Huawei Technologies Co., LTD. All rights reserved.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of Apache License, Version 2.0.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * Apache License, Version 2.0 for more details.
+ */
+package com.smn.response.sms;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+/**
+ * sms sign info data
+ *
+ * @author zhangyx
+ * @version 2.0.0
+ */
+public class SmsSignInfo {
+    @JsonProperty("sign_name")
+    private String signName;
+
+    @JsonProperty("create_time")
+    private String createTime;
+
+    @JsonProperty("sign_id")
+    private String signId;
+
+    @JsonProperty("reply")
+    private String reply;
+
+    @JsonProperty("status")
+    private int status;
+
+    public String getSignName() {
+        return signName;
+    }
+
+    public void setSignName(String signName) {
+        this.signName = signName;
+    }
+
+    public String getCreateTime() {
+        return createTime;
+    }
+
+    public void setCreateTime(String createTime) {
+        this.createTime = createTime;
+    }
+
+    public String getSignId() {
+        return signId;
+    }
+
+    public void setSignId(String signId) {
+        this.signId = signId;
+    }
+
+    public String getReply() {
+        return reply;
+    }
+
+    public void setReply(String reply) {
+        this.reply = reply;
+    }
+
+    public int getStatus() {
+        return status;
+    }
+
+    public void setStatus(int status) {
+        this.status = status;
+    }
+}
