+++ smn-sdk-java/src/main/java/com/smn/response/template/ListMessageTemplatesResponse.java
@@ -46,78 +46,3 @@ public class ListMessageTemplatesResponse extends AbstractResponse {
         this.messageTemplates = messageTemplates;
     }
 }
-
-/**
- * message template list info
- *
- * @author zhangyx
- * @version 2.0.0
- */
-class MessageTemplateListInfo {
-
-    @JsonProperty("message_template_name")
-    private String messageTemplateName;
-
-    @JsonProperty("create_time")
-    private String createTime;
-
-    @JsonProperty("update_time")
-    private String updateTime;
-
-    @JsonProperty("message_template_id")
-    private String messageTemplateId;
-
-    @JsonProperty("tag_names")
-    private List<String> tagNames;
-
-    @JsonProperty("protocol")
-    private String protocol;
-
-    public String getMessageTemplateName() {
-        return messageTemplateName;
-    }
-
-    public void setMessageTemplateName(String messageTemplateName) {
-        this.messageTemplateName = messageTemplateName;
-    }
-
-    public String getCreateTime() {
-        return createTime;
-    }
-
-    public void setCreateTime(String createTime) {
-        this.createTime = createTime;
-    }
-
-    public String getUpdateTime() {
-        return updateTime;
-    }
-
-    public void setUpdateTime(String updateTime) {
-        this.updateTime = updateTime;
-    }
-
-    public String getMessageTemplateId() {
-        return messageTemplateId;
-    }
-
-    public void setMessageTemplateId(String messageTemplateId) {
-        this.messageTemplateId = messageTemplateId;
-    }
-
-    public List<String> getTagNames() {
-        return tagNames;
-    }
-
-    public void setTagNames(List<String> tagNames) {
-        this.tagNames = tagNames;
-    }
-
-    public String getProtocol() {
-        return protocol;
-    }
-
-    public void setProtocol(String protocol) {
-        this.protocol = protocol;
-    }
-}
