+++ smn-sdk-java/src/main/java/com/smn/response/topic/ListTopicsResponse.java
@@ -9,7 +9,6 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * Apache License, Version 2.0 for more details.
  */
-
 package com.smn.response.topic;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -45,58 +44,5 @@ public class ListTopicsResponse extends AbstractResponse {
 
     public void setTopics(List<TopicInfo> topics) {
         this.topics = topics;
-    }
-}
-
-/**
- * topic info
- *
- * @author zhangyx
- * @version 2.0.0
- */
-class TopicInfo {
-
-    @JsonProperty("name")
-    private String name;
-
-    @JsonProperty("topic_urn")
-    private String topicUrn;
-
-    @JsonProperty("display_name")
-    private String displayName;
-
-    @JsonProperty("push_policy")
-    private String pushPolicy;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getTopicUrn() {
-        return topicUrn;
-    }
-
-    public void setTopicUrn(String topicUrn) {
-        this.topicUrn = topicUrn;
-    }
-
-    public String getDisplayName() {
-        return displayName;
-    }
-
-    public void setDisplayName(String displayName) {
-        this.displayName = displayName;
-    }
-
-    public String getPushPolicy() {
-        return pushPolicy;
-    }
-
-    public void setPushPolicy(String pushPolicy) {
-        this.pushPolicy = pushPolicy;
     }
 }
